requirements.txt
asgiref==3.7.2
boto3==1.33.13
botocore==1.33.13
Django==3.2.25
django-storages==1.14.4
gunicorn==23.0.0
importlib-metadata==6.7.0
jmespath==1.0.1
packaging==24.0
Pillow==9.5.0
pysqlite3==0.5.4
python-dateutil==2.9.0.post0
pytz==2024.2
s3transfer==0.8.2
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.7.1
urllib3==1.26.20
zipp==3.15.0
